fix: refactor mock file directory.
